Move ResourceMetadata from ClientXdsClient to XdsClient
